everything runs and can close without validation warnings other than the semaphore.  resizing still doesn't work.  Next step is to get the parts of the engine to demonstrate things.

+++ attachments/simple_engine/engine.h
@@ -180,7 +180,7 @@ private:
 
     // Delta time calculation
     float deltaTime = 0.0f;
-    float lastFrameTime = 0.0f;
+    uint64_t lastFrameTime = 0;
 
     /**
      * @brief Update the engine state.
